chore: rebuild

+++ qcs-api-client-openapi/internal/src/apis/account_api.rs
@@ -1,7 +1,7 @@
 /*
  * Rigetti QCS API
  *
- * # Introduction  This is the documentation for the Rigetti QCS HTTP API.  You can find out more about Rigetti at [https://rigetti.com](https://rigetti.com), and also interact with QCS via the web at [https://qcs.rigetti.com](https://qcs.rigetti.com).  This API is documented in **OpenAPI format** and so is compatible with the dozens of language-specific client generators available [here](https://github.com/OpenAPITools/openapi-generator) and elsewhere on the web.  # Principles  This API follows REST design principles where appropriate, and otherwise an HTTP RPC paradigm. We adhere to the Google [API Improvement Proposals](https://google.aip.dev/general) where reasonable to provide a consistent, intuitive developer experience. HTTP response codes match their specifications, and error messages fit a common format.  # Authentication  All access to the QCS API requires OAuth2 authentication provided by Okta. You can request access [here](https://www.rigetti.com/get-quantum). Once you have a user account, you can download your access token from QCS [here](https://qcs.rigetti.com/auth/token).   That access token is valid for 24 hours after issuance. The value of `access_token` within the JSON file is the token used for authentication (don't use the entire JSON file).  Authenticate requests using the `Authorization` header and a `Bearer` prefix:  ``` curl --header \"Authorization: Bearer eyJraW...Iow\" ```  # Quantum Processor Access  Access to the quantum processors themselves is not yet provided directly by this HTTP API, but is instead performed over ZeroMQ/[rpcq](https://gitlab.com/rigetti/rpcq). Until that changes, we suggest using [pyquil](https://gitlab.com/rigetti/pyquil) to build and execute quantum programs via the Legacy API.  # Legacy API  Our legacy HTTP API remains accessible at https://forest-server.qcs.rigetti.com, and it shares a source of truth with this API's services. You can use either service with the same user account and means of authentication. We strongly recommend using the API documented here, as the legacy API is on the path to deprecation.
+ * # Introduction  This is the documentation for the Rigetti QCS HTTP API.  You can find out more about Rigetti at [https://rigetti.com](https://rigetti.com), and also interact with QCS via the web at [https://qcs.rigetti.com](https://qcs.rigetti.com).  This API is documented in **OpenAPI format** and so is compatible with the dozens of language-specific client generators available [here](https://github.com/OpenAPITools/openapi-generator) and elsewhere on the web.  # Principles  This API follows REST design principles where appropriate, and otherwise an HTTP RPC paradigm. We adhere to the Google [API Improvement Proposals](https://google.aip.dev/general) where reasonable to provide a consistent, intuitive developer experience. HTTP response codes match their specifications, and error messages fit a common format.  # Authentication  All access to the QCS API requires OAuth2 authentication provided by Okta. You can request access [here](https://www.rigetti.com/get-quantum). Once you have a user account, you can download your access token from QCS [here](https://qcs.rigetti.com/auth/token).   That access token is valid for 24 hours after issuance. The value of `access_token` within the JSON file is the token used for authentication (don't use the entire JSON file).  Authenticate requests using the `Authorization` header and a `Bearer` prefix:  ``` curl --header \"Authorization: Bearer eyJraW...Iow\" ```  # Quantum Processor Access  Access to the quantum processors themselves is not yet provided directly by this HTTP API, but is instead performed over ZeroMQ/[rpcq](https://github.com/rigetti/rpcq). Until that changes, we suggest using [pyquil](https://github.com/rigetti/pyquil) to build and execute quantum programs via the Legacy API.  # Legacy API  Our legacy HTTP API remains accessible at https://forest-server.qcs.rigetti.com, and it shares a source of truth with this API's services. You can use either service with the same user account and means of authentication. We strongly recommend using the API documented here, as the legacy API is on the path to deprecation.
  *
  * The version of the OpenAPI document: 2020-07-31
  * Contact: support@rigetti.com
@@ -1171,7 +1171,7 @@ pub async fn internal_create_user_metered_usage_record(
 }
 async fn internal_delete_event_billing_price_inner(
     configuration: &configuration::Configuration,
-    event_billing_price_id: i32,
+    event_billing_price_id: i64,
 ) -> Result<crate::models::EventBillingPrice, Error<InternalDeleteEventBillingPriceError>> {
     let local_var_configuration = configuration;
 
@@ -1213,7 +1213,7 @@ async fn internal_delete_event_billing_price_inner(
 /// Delete an event billing price.
 pub async fn internal_delete_event_billing_price(
     configuration: &configuration::Configuration,
-    event_billing_price_id: i32,
+    event_billing_price_id: i64,
 ) -> Result<crate::models::EventBillingPrice, Error<InternalDeleteEventBillingPriceError>> {
     match internal_delete_event_billing_price_inner(configuration, event_billing_price_id.clone())
         .await
@@ -1436,7 +1436,7 @@ async fn internal_list_event_billing_prices_inner(
     configuration: &configuration::Configuration,
     filter: Option<&str>,
     order: Option<&str>,
-    page_size: Option<i32>,
+    page_size: Option<i64>,
     page_token: Option<&str>,
     show_deleted: Option<&str>,
 ) -> Result<crate::models::ListEventBillingPricesResponse, Error<InternalListEventBillingPricesError>>
@@ -1503,7 +1503,7 @@ pub async fn internal_list_event_billing_prices(
     configuration: &configuration::Configuration,
     filter: Option<&str>,
     order: Option<&str>,
-    page_size: Option<i32>,
+    page_size: Option<i64>,
     page_token: Option<&str>,
     show_deleted: Option<&str>,
 ) -> Result<crate::models::ListEventBillingPricesResponse, Error<InternalListEventBillingPricesError>>
@@ -1538,7 +1538,7 @@ pub async fn internal_list_event_billing_prices(
 }
 async fn internal_list_groups_inner(
     configuration: &configuration::Configuration,
-    page_size: Option<i32>,
+    page_size: Option<i64>,
     page_token: Option<&str>,
 ) -> Result<crate::models::ListGroupsResponse, Error<InternalListGroupsError>> {
     let local_var_configuration = configuration;
@@ -1589,7 +1589,7 @@ async fn internal_list_groups_inner(
 /// List all groups available on QCS.
 pub async fn internal_list_groups(
     configuration: &configuration::Configuration,
-    page_size: Option<i32>,
+    page_size: Option<i64>,
     page_token: Option<&str>,
 ) -> Result<crate::models::ListGroupsResponse, Error<InternalListGroupsError>> {
     match internal_list_groups_inner(configuration, page_size.clone(), page_token.clone()).await {
@@ -1607,7 +1607,7 @@ async fn internal_list_users_inner(
     configuration: &configuration::Configuration,
     filter: &str,
     order: Option<&str>,
-    page_size: Option<i32>,
+    page_size: Option<i64>,
     page_token: Option<&str>,
 ) -> Result<crate::models::InternalListUsersResponse, Error<InternalListUsersError>> {
     let local_var_configuration = configuration;
@@ -1665,7 +1665,7 @@ pub async fn internal_list_users(
     configuration: &configuration::Configuration,
     filter: &str,
     order: Option<&str>,
-    page_size: Option<i32>,
+    page_size: Option<i64>,
     page_token: Option<&str>,
 ) -> Result<crate::models::InternalListUsersResponse, Error<InternalListUsersError>> {
     match internal_list_users_inner(
@@ -1689,7 +1689,7 @@ pub async fn internal_list_users(
 }
 async fn internal_update_event_billing_price_inner(
     configuration: &configuration::Configuration,
-    event_billing_price_id: i32,
+    event_billing_price_id: i64,
     internal_update_event_billing_price_request: crate::models::InternalUpdateEventBillingPriceRequest,
 ) -> Result<crate::models::EventBillingPrice, Error<InternalUpdateEventBillingPriceError>> {
     let local_var_configuration = configuration;
@@ -1735,7 +1735,7 @@ async fn internal_update_event_billing_price_inner(
 /// Update an event billing price. This will also set the Stripe price in the server cache.
 pub async fn internal_update_event_billing_price(
     configuration: &configuration::Configuration,
-    event_billing_price_id: i32,
+    event_billing_price_id: i64,
     internal_update_event_billing_price_request: crate::models::InternalUpdateEventBillingPriceRequest,
 ) -> Result<crate::models::EventBillingPrice, Error<InternalUpdateEventBillingPriceError>> {
     match internal_update_event_billing_price_inner(
@@ -1823,7 +1823,7 @@ async fn list_group_billing_invoice_lines_inner(
     group_name: &str,
     billing_invoice_id: &str,
     page_token: Option<&str>,
-    page_size: Option<i32>,
+    page_size: Option<i64>,
 ) -> Result<
     crate::models::ListAccountBillingInvoiceLinesResponse,
     Error<ListGroupBillingInvoiceLinesError>,
@@ -1881,7 +1881,7 @@ pub async fn list_group_billing_invoice_lines(
     group_name: &str,
     billing_invoice_id: &str,
     page_token: Option<&str>,
-    page_size: Option<i32>,
+    page_size: Option<i64>,
 ) -> Result<
     crate::models::ListAccountBillingInvoiceLinesResponse,
     Error<ListGroupBillingInvoiceLinesError>,
@@ -1916,7 +1916,7 @@ async fn list_group_billing_invoices_inner(
     configuration: &configuration::Configuration,
     group_name: &str,
     page_token: Option<&str>,
-    page_size: Option<i32>,
+    page_size: Option<i64>,
 ) -> Result<crate::models::ListAccountBillingInvoicesResponse, Error<ListGroupBillingInvoicesError>>
 {
     let local_var_configuration = configuration;
@@ -1970,7 +1970,7 @@ pub async fn list_group_billing_invoices(
     configuration: &configuration::Configuration,
     group_name: &str,
     page_token: Option<&str>,
-    page_size: Option<i32>,
+    page_size: Option<i64>,
 ) -> Result<crate::models::ListAccountBillingInvoicesResponse, Error<ListGroupBillingInvoicesError>>
 {
     match list_group_billing_invoices_inner(
@@ -1996,7 +1996,7 @@ async fn list_group_upcoming_billing_invoice_lines_inner(
     configuration: &configuration::Configuration,
     group_name: &str,
     page_token: Option<&str>,
-    page_size: Option<i32>,
+    page_size: Option<i64>,
 ) -> Result<
     crate::models::ListAccountBillingInvoiceLinesResponse,
     Error<ListGroupUpcomingBillingInvoiceLinesError>,
@@ -2052,7 +2052,7 @@ pub async fn list_group_upcoming_billing_invoice_lines(
     configuration: &configuration::Configuration,
     group_name: &str,
     page_token: Option<&str>,
-    page_size: Option<i32>,
+    page_size: Option<i64>,
 ) -> Result<
     crate::models::ListAccountBillingInvoiceLinesResponse,
     Error<ListGroupUpcomingBillingInvoiceLinesError>,
@@ -2084,7 +2084,7 @@ pub async fn list_group_upcoming_billing_invoice_lines(
 async fn list_group_users_inner(
     configuration: &configuration::Configuration,
     group_name: &str,
-    page_size: Option<i32>,
+    page_size: Option<i64>,
     page_token: Option<&str>,
 ) -> Result<crate::models::ListGroupUsersResponse, Error<ListGroupUsersError>> {
     let local_var_configuration = configuration;
@@ -2137,7 +2137,7 @@ async fn list_group_users_inner(
 pub async fn list_group_users(
     configuration: &configuration::Configuration,
     group_name: &str,
-    page_size: Option<i32>,
+    page_size: Option<i64>,
     page_token: Option<&str>,
 ) -> Result<crate::models::ListGroupUsersResponse, Error<ListGroupUsersError>> {
     match list_group_users_inner(
@@ -2163,7 +2163,7 @@ async fn list_user_billing_invoice_lines_inner(
     user_id: &str,
     billing_invoice_id: &str,
     page_token: Option<&str>,
-    page_size: Option<i32>,
+    page_size: Option<i64>,
 ) -> Result<
     crate::models::ListAccountBillingInvoiceLinesResponse,
     Error<ListUserBillingInvoiceLinesError>,
@@ -2221,7 +2221,7 @@ pub async fn list_user_billing_invoice_lines(
     user_id: &str,
     billing_invoice_id: &str,
     page_token: Option<&str>,
-    page_size: Option<i32>,
+    page_size: Option<i64>,
 ) -> Result<
     crate::models::ListAccountBillingInvoiceLinesResponse,
     Error<ListUserBillingInvoiceLinesError>,
@@ -2256,7 +2256,7 @@ async fn list_user_billing_invoices_inner(
     configuration: &configuration::Configuration,
     user_id: &str,
     page_token: Option<&str>,
-    page_size: Option<i32>,
+    page_size: Option<i64>,
 ) -> Result<crate::models::ListAccountBillingInvoicesResponse, Error<ListUserBillingInvoicesError>>
 {
     let local_var_configuration = configuration;
@@ -2310,7 +2310,7 @@ pub async fn list_user_billing_invoices(
     configuration: &configuration::Configuration,
     user_id: &str,
     page_token: Option<&str>,
-    page_size: Option<i32>,
+    page_size: Option<i64>,
 ) -> Result<crate::models::ListAccountBillingInvoicesResponse, Error<ListUserBillingInvoicesError>>
 {
     match list_user_billing_invoices_inner(
@@ -2335,7 +2335,7 @@ pub async fn list_user_billing_invoices(
 async fn list_user_groups_inner(
     configuration: &configuration::Configuration,
     user_id: &str,
-    page_size: Option<i32>,
+    page_size: Option<i64>,
     page_token: Option<&str>,
 ) -> Result<crate::models::ListGroupsResponse, Error<ListUserGroupsError>> {
     let local_var_configuration = configuration;
@@ -2388,7 +2388,7 @@ async fn list_user_groups_inner(
 pub async fn list_user_groups(
     configuration: &configuration::Configuration,
     user_id: &str,
-    page_size: Option<i32>,
+    page_size: Option<i64>,
     page_token: Option<&str>,
 ) -> Result<crate::models::ListGroupsResponse, Error<ListUserGroupsError>> {
     match list_user_groups_inner(
@@ -2413,7 +2413,7 @@ async fn list_user_upcoming_billing_invoice_lines_inner(
     configuration: &configuration::Configuration,
     user_id: &str,
     page_token: Option<&str>,
-    page_size: Option<i32>,
+    page_size: Option<i64>,
 ) -> Result<
     crate::models::ListAccountBillingInvoiceLinesResponse,
     Error<ListUserUpcomingBillingInvoiceLinesError>,
@@ -2469,7 +2469,7 @@ pub async fn list_user_upcoming_billing_invoice_lines(
     configuration: &configuration::Configuration,
     user_id: &str,
     page_token: Option<&str>,
-    page_size: Option<i32>,
+    page_size: Option<i64>,
 ) -> Result<
     crate::models::ListAccountBillingInvoiceLinesResponse,
     Error<ListUserUpcomingBillingInvoiceLinesError>,
